attend to comments, refactor agent folder, minor fixes
