--- perfkitbenchmarker/data/docker/agentic/adk-agent/gke_performance_agent/__init__.py
@@ -1,2 +1,0 @@
-# GKE Performance Agent package
-from . import agent
