+++ perfkitbenchmarker/data/docker/agentic/adk-agent/performance_agent/__init__.py
@@ -1,0 +1,2 @@
+# Performance Agent package
+from . import agent
