+++ perfkitbenchmarker/data/docker/agentic/adk-agent/requirements.txt
@@ -1,4 +1,4 @@
-# Requirements for GKE Performance Agent
+# Requirements for Performance Agent
 google-adk[gke,extensions]==1.34.1
 k8s-agent-sandbox==0.4.6
 kubernetes>=36.0.1  # Fix: v36.0.0 has auth key mismatch bug (PR #2585)
